refactor: extract bookmark sync logic into dedicated hook
Extract complex bookmark synchronization logic from page.tsx into
a dedicated `use-bookmark-sync.ts` hook for better separation of
concerns and reusability.

Features extracted:
- Cross-tab bookmark synchronization using storage events
- Automatic bookmark status syncing with domain results
- Event listeners for bookmark changes
- State management for bookmark sync tracking

This significantly reduces page.tsx complexity while maintaining
all bookmark sync functionality including cross-tab communication.
The hook is now reusable and properly encapsulated.

Implements:
- #49 Extract bookmark sync logic into dedicated hook

Completes homepage refactor epic #50
